test: simplify CMake integration fixtures
